‼️ Switch from `virtualenv` to  `venv`
Adapt the `VirtualenvProject` project class to rely on using `venv` instead of
`virtualenv`, renaming it to `VenvProject` in the process.

The `venv` package is part of the standard library, whereas `virtualenv` needs
to be installed separtely. This has the clear advantage that `venv` does not
need to be installed. Although `virtualenv` does have some upsides as well,
none of them are particularly relevant for `aiida-project`.