⬆️ Dependencies: update `pydantic` dependency
The release of `typing_extensions` version `4.6.0` highlighted a bug in
`pydantic` related to the `Literal` type for Python versions 3.8 and 3.9.
This was fixed in `pydantic` version `1.10.8`.

Here the version of `pydantic` is set to be a compatible release with `1.10`
but at least `1.10.8`.

Additionally, other dependencies are set to a compatible _minor_ release.