✨ Allow plugin installs from GitHub repository
When creating a new project with the `create` command, the `--plugin` option
now also accepts GitHub repository URLs. Such a repository will be automatically
cloned into the `git` subdirectory of the project directory, which has been
renamed from `code` to avoid confusion with the subdirectory of `setup` where
the AiiDA code YAML files should be stored.