🐛 Fix `append_deactivate_text` for `venv`
From Python 3.10, the `VIRTUAL_ENV_PROMPT` environment variable is also
set and unset in the `activate` script. Since our current (obviously rather
fragile) approach to adding text to the `deactivate` function in the `activate`
script relies on replacing `VIRTUAL_ENV`, this would break and give the
following error:

```
deactivate:31: command not found: _PROMPT
```

when activating or deactivating the environment.

Here we switch to replacing the `deactivate () {` line in the `activate` script
that. That should be more robust unless the function definition syntax is
changed.