✨ Add `fish` support
The only shells supported currently are `zsh` and `bash`. The friendly interactive
shell (`fish`) is also popular among developers and especially the sexy German ones.

Here we add support for `fish` by creating the `Shell` `BaseModel` that has four fields:

1. `config_file`: Path to shell configuration relative to home directory.
2. `init_lines`: Lines to add to the shell configuration files when using
    `aiida-project init`.
3. `activate`: AiiDA-specific lines to add to the environment's activate script.
4. `deactivate`: AiiDA-specific lines to add to the environment's deactivate script.

A `Shell` instance can be loaded from a string using the `load_shell` function, where
the field are populated from a single YAML file that contains the values of the fields
for each shell.

The `init` and `create` commands are updated to use the `Shell` instead of relying
on `bash`/`zsh` specific strings. Similarly, the `VenvProject` is adapted to
accommodate for the various shell options.

Co-authored-by: Julian Geiger <julian.geiger@epfl.ch>