👌 CLI: use `sys.exit` instead of `return` for errors
Currently the `create` command exits via `return` on failures (e.g. when the Python
binary cannot be found). This yields an exit status of 0, falsely signalling success:

    ❯ aiida-project create tmp --python 4.0; echo $status
    Error: Could not resolve path to Python binary.
    0

This makes it difficult to test the package in CI and is important when for example
a user relies on `&&` to only execute a subsequent command when the first one succeeds:

    ❯ aiida-project create tmp --python 4.0 && cda tmp

We replace `return` with `sys.exit`, using appropriate `os.EX_*` exit statuses.

Note that the `EX_NOTFOUND` exit status does not exist, despite being documented:

https://docs.python.org/3.12/library/os.html#os.EX_NOTFOUND

We cannot help but suspect this is an intented bug introduced by someone with a lovely
sense of irony.

Co-authored-by: Daniel Hollas <daniel.hollas@bristol.ac.uk>