📦 `dev`: switch `pre-commit` pin to lower bound `>=3.4`
`pre-commit` is a standalone tool rather than a project dependency,
so there is little reason to pin it save for a broken release,
which is unlikely to be a problem for long.
Pinning prevents getting updates and timely support for new Python versions
and old versions of pre-commit can fail to work with newer hooks.