👌 `create`: install packages together
The `--plugin` option of the `create` command allows the user to install additional
packages next to `aiida-core`. In the current approach these are installed one by one
after completing the installation of `aiida-core`. Moreover, if any of the package
installations failed, this would do so silently.

Here we improve the installation step by installing all packages in one go, which
makes it easier for the resolver to take into account all constraints. Additionally,
any errors are now caught and shown to the user.

This commit also removes the broken feature to clone packages from their GitHub
repository and install them locally. Note that installing packages from GitHub is still
possible using the typically `pip` syntax `git+`:

    aiida-project create test -p git+https://github.com/aiidateam/aiida-quantumespresso

The repo is simply no longer cloned to the `git` directory`. This directory is also
removed from the default project layout, since it was confusing to refer to it due to
the similarity with the `.git` directory.

Co-authored-by: Daniel Hollas <daniel.hollas@bristol.ac.uk>