👌 `VenvProject`: switch to `uv`
Switch to using `uv` to set up the virtual environment and install packages for
`VenvProject`. `uv` is added as a dependency, and the executable installed in the
Python environment where `aiida-project` is installed is used. In rare cases where this
does not work (e.g. installations in the global environment), we try to fall back on
looking for `uv` in the `PATH` via `which` and exit with an error in case that fails.

Although `uv` is still relatively new, it's currently used in most of our CI's, and
hence in case there is an issue we should discover it quite quickly. In case there is
still a need to have `pip` as a fallback, we can add a configuration option that allows
the user to do so.

Co-authored-by: Daniel Hollas <daniel.hollas@bristol.ac.uk>