👌 `create`: add additional checks and error handling
Currently the user can pass an already existing project or an empty string as the
project name of the `create` command. This can lead the user to potentially lose all
their `aiida-project` environments.

Here we check for both cases and exit with an error if true. Moreover, we also catch
any error raised by the `BaseProject.create()` method and echo the details to the user,
instead of just silently failing.

Co-authored-by: Daniel Hollas <daniel.hollas@bristol.ac.uk>